talloc: Make abi checks in release script
Make always sure the exports and signature files are up to date before
shipping a release.
(cherry picked from commit 8fb483b2964f1222c013df42ba8be544269da612)

(cherry picked from commit c751ceb929c5f025648aa4f3a5e746287eeee726)

+++ lib/talloc/script/release-script.sh
@@ -10,6 +10,16 @@ if [ ! -d "lib/talloc" ]; then
     exit 1
 fi
 
+# Check exports and signatures are up to date
+pushd lib/talloc
+./script/abi_checks.sh talloc talloc.h
+abicheck=$?
+popd
+if [ ! "$abicheck" = "0" ]; then
+    echo "ERROR: ABI Checks produced warnings!"
+    exit 1
+fi
+
 git clean -f -x -d lib/talloc
 git clean -f -x -d lib/replace
